Move dictionaries to their own module
